+++ aim/cli/digifeeds.py
@@ -6,9 +6,12 @@ import typer
 from typing_extensions import Annotated
 from aim.digifeeds.add_to_db import add_to_db as add_to_digifeeds_db
 from aim.digifeeds.list_barcodes_in_bucket import list_barcodes_in_bucket
-from aim.digifeeds.check_zephir import check_zephir as check_zephir_for_barcode
+
+# from aim.digifeeds.check_zephir import check_zephir as check_zephir_for_barcode
 from aim.digifeeds.move_to_pickup import move_to_pickup as move_volume_to_pickup
 from aim.digifeeds.database import models, main
+from aim.digifeeds import functions
+
 import json
 import sys
 
@@ -57,7 +60,7 @@ def check_zephir(
     """
 
     print(f"Checking Zephir for {barcode}")
-    item = check_zephir_for_barcode(barcode)
+    item = functions.check_zephir(barcode)
     if item:
         print(f"{barcode} is in Zephir")
     else:
