--- docs/api/aim.digifeeds.check_zephir.rst
@@ -1,7 +1,0 @@
-aim.digifeeds.check\_zephir module
-==================================
-
-.. automodule:: aim.digifeeds.check_zephir
-   :members:
-   :undoc-members:
-   :show-inheritance:
