+++ docs/api/aim.digifeeds.functions.rst
@@ -1,0 +1,7 @@
+aim.digifeeds.functions module
+==============================
+
+.. automodule:: aim.digifeeds.functions
+   :members:
+   :undoc-members:
+   :show-inheritance:
