+++ docs/api/aim.digifeeds.rst
@@ -22,8 +22,8 @@ Submodules
 
    aim.digifeeds.add_to_db
    aim.digifeeds.alma_client
-   aim.digifeeds.check_zephir
    aim.digifeeds.db_client
+   aim.digifeeds.functions
    aim.digifeeds.item
    aim.digifeeds.list_barcodes_in_bucket
    aim.digifeeds.move_to_pickup
